Reference spike for upgrading conduit version
This is a reference spike only. The Acquire implementations
are *probably* fine, but I was unable to implement a conduit
test because TestDB on this branch cannot implement UNLiftIO.
Furthermore, it's not clear whether we want to provide a
MonadUnliftIO instance for OrvilleT or not yet.

On top of that, I had to comment out redundant Monoid imports
to get this code compile in the development environment, but
those imports are not redundant on older versions of GHC.